FindProjectPage

+++ src/main/java/org/ayty/hatcher/api/v1/project/service/FindProjectByIdService.java
@@ -7,5 +7,5 @@ import java.util.Optional;
 @FunctionalInterface
 public interface FindProjectByIdService {
 
-    Optional<Project> findById(Integer id);
+    Optional<Project> findById(Long id);
 }
